Inserts inbound notifications into RDB (primative implementation).
Next step:
Refactor code into more logical blocks.
Use correct datatypes in RDB (esspecially the timestamps)
Implement agnostic primary keys.
Think about adding air-time as a field to geyser.

Design note - RDB, Jetty vs Tomcat:
If Tomcat was used as the servlet container, then database access can be
achived using the Tomcat database connection pool. In such a case you
don't have to worry about opening and closing the connections. However,
in this case a simple Jetty server is used as the servlet container and
therefore you have to manage the database connection manually. As far as
I understand, a database transaction is only finilized when the
connection is closed. Therefore in this case the connection will be open
and closed with each new geyser entry.